Add files via upload
In the above the change in utility against the change in the amount of a and b is shown, for income i=1000
price of a, pa=10
price of b, pb=20
weight of a on utility = 1
weight of b on utility = 3
the income relation will be given as 10a + 20b = 1000
utility relation as u = a + 3b (linear) and u = ab^3 (multiplicative)
Observing the multiplicative utility graph, the utility is 0 when either of the amount is 0.
 As the amount of a increases to a single unit, utility jumps rapidly as most of the income is going to b which has a cubic effect on utility. 
In case of b increasing from 0, the utility increases slowly as most of the income is going to a, which has a lower weight on utility. 
in linear case, the slope of b is higher as its unit increase provides more change in utility. as its price is higher it also reaches its peak quickly.